Populate Layers Phase 1 (see description)
I've separated the first phase of the function into:
a) get_vector_tile_layers, which handles the safe retrieval of the title objects, and,
b) no_maptiler_layers found, which stores the error message display code. This was being repeated multiple times in separate places and should've had its own function in the first place.